feat: move utils to single file
Will help for testing and managing properly
