+++ src/utils/hasProp.js
@@ -1,0 +1,9 @@
+function hasProp(obj, prop) {
+  if (!obj) return false;
+  if (Array.isArray(obj)) {
+    return obj.includes(prop);
+  }
+  return {}.hasOwnProperty.call(obj, prop);
+}
+
+module.exports = hasProp;
