+++ src/utils/index.js
@@ -1,28 +1,32 @@
+const extendTransactionsWithMetadata = require('./extendTransactionsWithMetadata');
+const categorizeTransactions = require('./categorizeTransactions');
+const filterTransactions = require('./filterTransactions');
+const { hash, doubleSha256, sha256 } = require('./crypto');
 const { varIntSizeBytesFromLength } = require('./varInt');
-const {
-  dashToDuffs,
-  duffsToDash,
-  getBytesOf,
-  hasProp,
-  hasMethod,
-} = require('./utils');
+const feeCalculation = require('./feeCalculation');
+const coinSelection = require('./coinSelection');
+const fundWallet = require('./fundWallet');
+const dashToDuffs = require('./dashToDuffs');
+const duffsToDash = require('./duffsToDash');
+const getBytesOf = require('./getBytesOf');
+const hasMethod = require('./hasMethod');
+const hasProp = require('./hasProp');
+const mnemonic = require('./mnemonic');
+const is = require('./is');
+
 const {
   generateNewMnemonic,
   mnemonicToHDPrivateKey,
   mnemonicToWalletId,
   seedToHDPrivateKey,
   mnemonicToSeed,
-} = require('./mnemonic');
-const is = require('./is');
-const coinSelection = require('./coinSelection');
-const feeCalculation = require('./feeCalculation');
-const { hash, doubleSha256, sha256 } = require('./crypto');
-
-const fundWallet = require('./fundWallet');
+} = mnemonic;
 
 module.exports = {
+  categorizeTransactions,
   dashToDuffs,
   duffsToDash,
+  extendTransactionsWithMetadata,
   generateNewMnemonic,
   mnemonicToHDPrivateKey,
   mnemonicToWalletId,
@@ -31,6 +35,7 @@ module.exports = {
   is,
   coinSelection,
   feeCalculation,
+  filterTransactions,
   varIntSizeBytesFromLength,
   getBytesOf,
   hash,
